backend_api/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
requests
aiohttp
python-multipart